+++ CHANGELOG.md
@@ -4,6 +4,8 @@ NCS Navigator MDES Module history
 0.3.2
 -----
 
+- Consider element `minOccurs` when determining variable
+  requiredness (#2).
 - Added disposition codes to the specification. Requires a YAML file 
   with all the disposition code values for the MDES version.